backend/pyproject.toml
[project]
name = "pfizer-brand-planning"
version = "0.1.0"
description = "Pfizer AI Brand Planning Prototype"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.1",
    "langgraph>=0.2.50",
    "langchain>=0.3.7",
    "langchain-openai>=0.2.9",
    "langsmith>=0.1.145",
    "supabase>=2.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "instructor>=1.6.4",
    "openai>=1.54.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
]
